Rename models module to data_structures
